==> 2_RNAseq_viz.Rmd <==
---
title: "Visualisation of DESeq2 results"
description: "DEG analysis based on DESeq2 and GSEA"
principal investigator: "Joaquín de Navascués"
researcher: "Aleix Puig, modified by Joaquín de Navascués"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
    theme: readable
    df_print: paged
---
```{r set-publication-theme, echo=FALSE, cache=FALSE}
ggplot2::theme_set(ggpubr::theme_pubr(base_size=10))
```

```{r setup, echo = FALSE, cache = FALSE}
knitr::opts_chunk$set(dev = c('png', 'cairo_pdf'), 
                      fig.align = 'center', fig.height = 5, fig.width = 8.5, 
                      pdf.options(encoding = "ISOLatin9.enc"),
                      fig.path='integration/figures/', warning=FALSE, message=FALSE)
```

```{r, include=FALSE}
# data
library(tibble)
# graphics
library(ggplot2)
library(ggthemes)       # italics in the figures
library(ggtext)         # italics in the figures
library(pheatmap)       # pretty heatmap
library(scico)          # scientific colormaps
library(ggrepel)        # to avoid overlapping text
# convenience
library(here);
setwd(here())
```

# 1 Principal Components Analysis

## Using the raw counts

```{r}
# get the data (fresh session)
targets <- readRDS('output/targets.RDS')
tpmNormalisedCounts <- readRDS('output/tpmNormalisedCounts.RDS')
normalisedCounts <- readRDS('output/normalisedCounts.RDS')
rawCounts <- readRDS('output/rawCounts.RDS')
```

Let us see what the variance looks like in `rawCounts`:
```{r}
var_rawCounts <- rowVars(as.matrix(rawCounts), rm.na=TRUE)
var_rawCounts <- var_rawCounts [ var_rawCounts < max(var_rawCounts) ]
avg_rawCounts <- rowMeans(as.matrix(rawCounts))
avg_rawCounts <- avg_rawCounts [ avg_rawCounts < max(avg_rawCounts) ]
df <- data.frame(Variance=var_rawCounts, Mean=avg_rawCounts)
p1 <- ggplot(df, aes(x=Variance, y=Mean)) +
  geom_point(alpha=0.2)
p2 <- ggplot(df, aes(x=Variance, y=Mean)) +
  geom_point(alpha=0.2) + xlim(0, 500) + ylim(0, 100) +
  theme(panel.background = element_rect(fill='grey90'))
p1 + annotation_custom(ggplotGrob(p2),
                       xmin=3.5e10, xmax=6e10,
                       ymin=1e2, ymax=8e5)
```

It does not look like it is necessary to filter for variance, so we can calculate the principal components straight away:
```{r}
pca <- prcomp(t(rawCounts), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
```

Plot:
```{r}
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(raw counts)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=1.2, force_pull=-0.02, 
                  box.padding=0.5, point.padding=0.5) +
  lims(x= c(-115, 175), y = c(-70, 70)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_text(hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```

## Using normalised counts

Let's do this again with the `normalisedCounts` and the `tpmNormalisedCounts`:
```{r}
pca <- prcomp(t(normalisedCounts), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(DESeq2-normalised counts)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=1.2, force_pull=-0.02, 
                  box.padding=0.5, point.padding=0.5) +
  lims(x= c(-115, 175), y = c(-80, 100)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_text(hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```

```{r}
pca <- prcomp(t(tpmNormalisedCounts), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(transcripts per million counts)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=1.2, force_pull=-0.02, 
                  box.padding=0.5, point.padding=0.5) +
  lims(x= c(-115, 175), y = c(-70, 70)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_text(hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```

The differences between these 3 are minimal, and the groupings are clearly dominated by biological replicates and batches.

## Using variance-stabilising transformed-counts, with and without batch correction

Now with the _variance-stabilising_-transformed counts (pseudocounts):
```{r, warning=FALSE}
vsd <- readRDS('output/vst_pseudocounts.RDS')
# remove rows with very little variability
# vsd_pca <- assay(vsd)[
#   rowSums(rawCounts[-1] != rawCounts[[2]], na.rm = TRUE) != 0, ]
var_vsd <- rowVars(assay(vsd), rm.na=TRUE)
#var_rawCounts <- var_rawCounts [ var_rawCounts < max(var_rawCounts) ]
avg_vsd <- rowMeans(assay(vsd))
#avg_rawCounts <- avg_rawCounts [ avg_rawCounts < max(avg_rawCounts) ]
df <- data.frame(Variance=var_vsd, Mean=avg_vsd)
p1 <- ggplot(df, aes(x=Variance, y=Mean)) +
  geom_point(alpha=0.2)
p2 <- ggplot(df, aes(x=Variance, y=Mean)) +
  geom_point(alpha=0.2) + xlim(0,0.1) + ylim(5,15) +
  theme(panel.background = element_rect(fill='grey90'))
p1 + annotation_custom(ggplotGrob(p2),
                       xmin=7, xmax=15,
                       ymin=10, ymax=25)
```

Some genes have very low variance (which is the whole point of the `vst` transformation), but none of them are zero.
There seems to be little point in drawing an arbitrary threshold, so I move on with this.

```{r}
pca <- prcomp(t(assay(vsd)), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(*vst* pseudocounts)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=1.2, force_pull=-0.02, 
                  box.padding=0.5, point.padding=0.5) +
  lims(x= c(-115, 175), y = c(-80, 80)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_markdown(hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```
```{r}
ggsave('PCA_vsd.pdf', plot = last_plot(), device = 'pdf',
       path = 'output', dpi = 300)
ggsave('PCA_vsd.svg', plot = last_plot(), device = 'svg',
       path = 'output', dpi = 300)
ggsave('PCA_vsd.png', plot = last_plot(), device = png,
       path = 'output', dpi = 300)
```

Again, very similar to the other 3, which is (marginally) reassuring.

Now, if we do the same with the **batch-corrected** VS-transformed data:
```{r}
vsd_bcorr <- readRDS('output/vst_pseudocounts_batchCorrected.RDS')
pca <- prcomp(t(assay(vsd_bcorr)), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(*vst* pseudocounts, batch-corrected)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=0.3, force_pull=1, 
                  box.padding=1, point.padding=0.5) +
  lims(x= c(-100, 150), y = c(-80, 80)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_markdown(hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```

```{r}
ggsave('PCA_vsd_bcorr.pdf', plot = last_plot(), device = 'pdf',
       path = 'output', dpi = 300)
ggsave('PCA_vsd_bcorr.svg', plot = last_plot(), device = 'svg',
       path = 'output', dpi = 300)
ggsave('PCA_vsd_bcorr.png', plot = last_plot(), device = png,
       path = 'output', dpi = 300)
```

## Using TPM counts, attempting a batch correction

Let us say we want to stubbornly stick to 'real-world' measurements (e.g. tpms).
We can do batch correction using `limma` on `tpmNormalisedCounts`, provided these are in log form.
So we do:
```{r}
# min(tpmNormalisedCounts) is 0 so we need to add 1 for log transform
tpm_log_bcorr <- limma::removeBatchEffect(
  log2(tpmNormalisedCounts+1),
  batch=vsd$batch,
  design=model.matrix(~condition, colData(vsd) )
)
# reverse log transform
tpm_bcorr <- 2^tpm_log_bcorr
# use in PCA
pca <- prcomp(t(tpm_bcorr), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(transcripts per million counts, batch-corrected)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=1.2, force_pull=-0.02, 
                  box.padding=0.5, point.padding=0.5) +
  #lims(x= c(-115, 175), y = c(-70, 70)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_text(hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```

It is not too bad, but it is not any less artificial that using pseudocounts from the moment we did the _log~2~(n+n~0~)_ transform, and the VS-transformed data worked better.
If we do not reverse the transform:

```{r}
pca <- prcomp(t(tpm_log_bcorr), center=TRUE, scale=TRUE)
scores <- data.frame(targets$sampleIDs, pca$x[,1:2])
summary(pca)
xtitle <- paste0('**PC1** (', round(summary(pca)$importance['Proportion of Variance','PC1']*100,1), ' %)')
ytitle <- paste0('**PC2** (', round(summary(pca)$importance['Proportion of Variance','PC2']*100,1), ' %)')
ptitle <- 'Principal component scores' 
stitle <- '(transcripts per million counts, *log~2~(n+n~0~)*-transformed and batch-corrected)'

ggplot(scores,
       aes(x = PC1, y = PC2,
           label=factor(targets$sampleIDs),
           colour=factor(targets$condition_md) )
       ) +
  # data
  geom_vline(xintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_hline(yintercept=0, linetype='dashed', colour='grey60', linewidth=0.5) +
  geom_point(size=4) +
  geom_point(size=2, colour='white', alpha=0.5) + 
  geom_text_repel(size=4, alpha=0.75,
                  max.overlaps=Inf,
                  seed=42,
                  force=1.2, force_pull=-0.02, 
                  box.padding=0.5, point.padding=0.5) +
  #lims(x= c(-115, 175), y = c(-70, 70)) +
  # decorations
  theme_minimal(base_size=16) +
  labs(x=xtitle,
       y=ytitle,
       title=ptitle,
       subtitle=stitle) +
  scale_colour_discrete(name="condition") +
  theme(legend.text = element_markdown(size=10),
        axis.title.x = element_markdown(),
        axis.title.y = element_markdown(),
        plot.title = element_text(hjust=0.5),
        plot.subtitle = element_markdown(size=12, hjust=0.5, color="grey40"),
        axis.text.x = element_text(color="grey60"),
        axis.text.y = element_text(color="grey60"),
        axis.ticks.x = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.y = element_line(linewidth=0.5, linetype = "solid", color="grey60"),
        axis.ticks.length=unit(-0.25, "cm"),
        panel.border = element_rect(linewidth=1, linetype = "solid", fill = NA),
        panel.grid = element_blank())
```
Quite similar. So I will stick to the `vst`-transformed pseudocounts for the paper.

# 2 Heatmap

## Using log~2~(TPM+1) with and without batch-correction
Do Z-scores on top of this. range of values is 0.2 - 10,000.
(or use the log2(TPM+1) transform, ranges -2 to 16)

```{r}
vsd <- assay(vst(dds))
Z <- t(scale(t(vsd)))
```














## Using pseudocounts with batch-correction
No need to do Z-scoring or any other normalisation here?? the range of values is ~ 5-22

(What would be the point of doing that *without* batch correction)

https://hbctraining.github.io/GCC-BOSC-2018/lessons/data_visualization.html


















Plot the tpm values

```{r}
relative_Significant_Genes <- sortedFinalData_named_KD %>% 
  filter(padj <= 0.01) %>%
  filter(log2FoldChange >= 1 | log2FoldChange <= -1) %>%
  dplyr::select(external_gene_name, starts_with("TPM")) %>%
  column_to_rownames(var="external_gene_name") %>%
  data.matrix()

#heatmap(heatmap_prep)


vsd <- assay(vst(dds))
Z <- t(scale(t(vsd)))

```
