==> 4_RNAseqVis_Sets.Rmd <==
---
title: "Visualisation of DESeq2 results with sets"
description: "DEG analysis based on DESeq2 and GSEA"
principal investigator: "Joaquín de Navascués"
researcher: "Aleix Puig, modified by Joaquín de Navascués"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: 'hide'
    theme: readable
    df_print: paged
---
```{r set-publication-theme, echo=FALSE, cache=FALSE}
ggplot2::theme_set(ggpubr::theme_pubr(base_size=10))
```

```{r setup, echo = FALSE, cache = FALSE}
knitr::opts_chunk$set(dev = 'png', 
                      fig.align = 'center', fig.height = 5, fig.width = 8.5, 
                      pdf.options(encoding = "ISOLatin9.enc"),
                      fig.path='integration/figures/', warning=FALSE, message=FALSE)
```

**Libraries/utils:**
```{r load_libraries, warning=FALSE}
if (!require("librarian")) install.packages("librarian")
# data
librarian::shelf(dplyr, purrr, stringr, reshape2, biomaRt, DescTools, santoku,
# graphics
UpSetR, ggVennDiagram, ggvenn, eulerr, RColorBrewer, cetcolor,
# convenience
here)

if(!exists("extract_regulated_sets", mode="function")) source("utils.R")
```

**Set working directory where the script is**
```{r setwd}
if (Sys.getenv("USER")=="JQ") {
  setwd("/Users/JQ/Documents/_CODE REPOS/GitHub/Da_RNAseq")
} else if (Sys.getenv("RSTUDIO")==1) {
  setwd( dirname(rstudioapi::getSourceEditorContext(id = NULL)$path) ) # gets what is in the editor
} else {
  setwd(here::here())
  d <- str_split(getwd(),'/')[[1]][length(str_split(getwd(),'/')[[1]])]
  if (d != 'Da_RNAseq') { stop(
    paste0("Could not set working directory automatically to where this",
           " script resides.\nPlease do `setwd()` manually"))
    }
}
getwd()
```

**Path to definitive images (outside repo):**
```{r define_dir2figs}
figdir <- paste0(c(head(str_split(getwd(),'/')[[1]],-1),
                   paste0(tail(str_split(getwd(),'/')[[1]],1), '_figures')),
                 collapse='/')
dir.create(figdir, showWarnings = FALSE)
```


# RNAseq results visualisation with sets


## 1 Getting ready


This gets us the DGE data from `DESeq2`, identified by FlyBase/Ensembl ID and gene symbol:
```{r load_DEG_data}
# experimental design and labels
targets <- readRDS('output/targets.RDS')
# DEG data
DaDaOE_deg <- readRDS('output/Control_vs_DaDaOE.RDS')
DaKD_deg <- readRDS('output/Control_vs_DaKD.RDS')
DaOE_deg <- readRDS('output/Control_vs_DaOE.RDS')
ScOE_deg <- readRDS('output/Control_vs_ScOE.RDS')
# gene IDs and symbols
ensembl = useEnsembl(biomart = "ENSEMBL_MART_ENSEMBL",
                     dataset="dmelanogaster_gene_ensembl",
                     host = "https://oct2022.archive.ensembl.org")
filters = listFilters(ensembl)       # defines filters for specific queries
attributes = listAttributes(ensembl) # defines features tro show
dlist <- getBM(attributes=c('ensembl_gene_id', 'external_gene_name'), mart = ensembl)
rownames(dlist) <- dlist$ensembl_gene_id
```

Now, for each condition, we will need to get the lists of misregulated genes for a given threshold of fold change:
```{r extract_regulated_sets, warning = FALSE}
list_of_degs <- list(DaKD_deg, DaOE_deg, DaDaOE_deg, ScOE_deg)
names_degs <- list(
  as.character(targets[targets$Condition=='DaKD','condition_md'][[1]]),
  as.character(targets[targets$Condition=='DaOE','condition_md'][[1]]),
  as.character(targets[targets$Condition=='DaDaOE','condition_md'][[1]]),
  as.character(targets[targets$Condition=='ScOE','condition_md'][[1]])
  )
regs <- extract_regulated_sets(list_of_degs, names_degs, fc_thresh=1.5)
head(regs,2)
```

Now we need to turn this into logicals:
```{r extract_up|down}
upreg   <- get_deg_logical(regs,'up')
downreg <- get_deg_logical(regs,'down')
```


## 2 Euler and Venn diagrams


### Euler diagrams using `eulerr`


Using instructions from the [vignette](https://cran.r-project.org/web/packages/eulerr/vignettes/introduction.html):
```{r euler_plots, results='hide'}
up_colours <- c(
  brewer.pal(9,'BrBG')[3],   # daRNAi   brown
  brewer.pal(9,'YlOrBr')[4], # daOVEX   yellow/orange
  brewer.pal(9,'Reds')[5],   # dadaOVEX red
  brewer.pal(9,'PuRd')[3]    # scOVEX   violet
)
down_colours <- c(
  brewer.pal(11,'PRGn')[8],  # daRNAi   green
  brewer.pal(11,'RdYlBu')[8], # daOVEX   pale blue
  brewer.pal(11,'RdBu')[9],  # dadaOVEX blue
  brewer.pal(9,'Purples')[5]     # scOVEX   purple
)
upfit <- euler(upreg, loss = 'region', loss_aggregator = 'max')
downfit <- euler(downreg, loss = 'region', loss_aggregator = 'max')
upplot <- plot(upfit,
               fill = up_colours,
               quantities = TRUE)
downplot <- plot(downfit,
                 fill = down_colours,
                 quantities = TRUE)
# save it
pdf(file=paste0(figdir,'/euler_up.pdf'),
     width=10, height=12)
upplot
dev.off()

pdf(file=paste0(figdir,'/euler_down.pdf'),
     width=10, height=12)
downplot
dev.off()

print(upplot)
print(downplot)
```

### Venn diagrams using `ggvenn`

We now need the first column to have the names of the elements in the set, so the `rownames` need become the first column:
```{r prepare_data_for_ggvenn, results='hide'}
upreg   <- cbind(data.frame(ensemblGeneID=rownames(upreg)), upreg)
downreg <- cbind(data.frame(ensemblGeneID=rownames(downreg)), downreg)
head(upreg,2)
```

Now both `upreg` and `downreg` are data.frames, which is what `ggvenn` wants as input.

Using instructions from the [vignette](https://github.com/yanlinlin82/ggvenn):
```{r plot_ggvenn_up, fig.width=18, warning=FALSE}
upvenn <- ggplot(upreg,
                 aes(A = `*da^RNAi^*`,
                     B = `*da*`,
                     C = `*da:da*`,
                     D = `*scute*`)) +
  geom_venn(show_percentage = FALSE) +
  theme_void() +
  coord_fixed()
upvenn
```
```{r prepare_data_for_ggvenn2, results='hide'}
downreg2 <- downreg
names(downreg2) <- c('ensemblGeneID', 'daKD', 'daOE', 'dadaOE', 'scOE')
```

```{r plot_ggvenn_down, fig.width=18, warning=FALSE}
downvenn <- ggplot(downreg2,
                   aes(A = `daKD`,
                       B = `daOE`,
                       C = `dadaOE`,
                       D = `scOE`)) +
  geom_venn(show_percentage = FALSE) +
  theme_void() +
  coord_fixed()
downvenn
```

This is, however, not very informative, visually. Maybe with `ggVennDiagram`?


### Venn diagrams using `ggVennDiagram`


From the CRAN [readme](https://cran.r-project.org/web/packages/ggVennDiagram/readme/README.html):

```{r plot_ggvenndiagram_upreg}
up <- list(`da RNAi`    = upreg[upreg$`*da^RNAi^*`,]$ensemblGeneID,
           `da`    = upreg[upreg$`*da*`,]$ensemblGeneID,
           `da:da` = upreg[upreg$`*da:da*`,]$ensemblGeneID,
           `scute` = upreg[upreg$`*scute*`,]$ensemblGeneID)
venn <- Venn(up)
data <- process_data(venn)
upvd <- ggplot() +
  # regions
  geom_sf(aes(fill=count), data = venn_region(data)) +
  geom_sf_text(aes(label=count), data = venn_region(data)) +
  scale_fill_gradientn(colours = c('grey50',cet_pal(n = 100, name = "cbd1", alpha = 1)[60:100])) +
  # edges
  geom_sf(size = 3, lty = "solid", color = "grey25", alpha=0.5, data = venn_setedge(data), show.legend = F) +
  # labels
  geom_sf_text(aes(label = name), data = venn_setlabel(data), fontface='bold.italic') +
  # space
  scale_x_continuous(expand = expansion(mult = .2)) +
  theme_void()
upvd
```
```{r plot_ggvenndiagram_downreg}
down <- list(`da RNAi`    = downreg[downreg$`*da^RNAi^*`,]$ensemblGeneID,
             `da`    = downreg[downreg$`*da*`,]$ensemblGeneID,
             `da:da` = downreg[downreg$`*da:da*`,]$ensemblGeneID,
             `scute` = downreg[downreg$`*scute*`,]$ensemblGeneID)
venn <- Venn(down)
data <- process_data(venn)
upvd <- ggplot() +
  # regions
  geom_sf(aes(fill=count), data = venn_region(data)) +
  geom_sf_text(aes(label=count), data = venn_region(data)) +
  scale_fill_gradientn(colours = c('grey50',cet_pal(n = 100, name = "cbd1", alpha = 1)[40:1])) +
  # edges
  geom_sf(size = 3, lty = "solid", color = "grey25", alpha=0.5, data = venn_setedge(data), show.legend = F) +
  # labels
  geom_sf_text(aes(label = name), data = venn_setlabel(data), fontface='bold.italic') +
  # space
  scale_x_continuous(expand = expansion(mult = .2)) +
  theme_void()
upvd
```

## 3 UpSet diagrams with `UpSetR`

```{r UpSet_plots_up}
# up
upsetplot <- upset(
  # data
  fromList(  up  ),
  mb.ratio = c(0.7, 0.3), # intersection bars, set bars
  queries = list(
    list(query = intersects, 
         params = list('da RNAi'),
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[2]],
         active = T),
    list(query = intersects, 
         params = list('da'),
         #color = cet_pal(n = 5, name = "cbd1", alpha = 1)[[4]],
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[2]],
         active = T),
    list(query = intersects, 
         params = list('da:da'),
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[2]],
         active = T),
    list(query = intersects, 
         params = list('scute'),
         #color = cet_pal(n = 5, name = "cbd1", alpha = 1)[[4]],
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[2]],
         active = T)
    ),
  # sets
  sets = names(up), keep.order= TRUE,
  sets.bar.color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[2]],
  # sets.bar.color = rep(c(cet_pal(n = 2, name = "cbd1", alpha = 1)[[2]],
  #                        cet_pal(n = 5, name = "cbd1", alpha = 1)[[4]]),
  #                      2),
  set_size.show = TRUE,
  set_size.numbers_size = 6,
  set_size.scale_max = (ceiling(max(sapply( up , length)) / 100)*100)+50, # analog to plyr::round_any
  # intersections
  main.bar.color = 'gray50',
  # c(rep("black", 6), rep('gray50', 8), rep("black", 6), rep('gray50', 7)),
  group.by = 'sets',
  order.by = "degree",
  # labels
  number.angles = 0,
  mainbar.y.label = "Condition Intersections",
  sets.x.label = "DEGs per condition", 
  text.scale = c(1.5, 1.3,  # c(ylab, yticks,
                 1.5, 1.3,  #   xlab, xticks,
                 1.3, 1.3), #   setlabs, barnumbers)
  # connections
  matrix.color = 'gray50',
  shade.color = cet_pal(n = 5, name = "cbd1", alpha = 1)[[4]],
  point.size = 3,
  line.size = 0.75
)

pdf(file=paste0(figdir,'/UpSet_plot_up_logfc2.pdf'),
    width=8, height=4.5)
upsetplot
grid.text("Set intersections for DEGs fc>4", x = 0.65, y=0.95, gp=gpar(fontsize=18)) #≥
dev.off()

upsetplot
grid.text("Set intersections for DEGs fc≥4", x = 0.65, y=0.95, gp=gpar(fontsize=18))
```

```{r UpSet_plots_down}
downsetplot <- upset(
  # data
  fromList(  down  ),
  mb.ratio = c(0.7, 0.3), # intersection bars, set bars
  queries = list(
    list(query = intersects, 
         params = list('da RNAi'),
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[1]],
         active = T),
    list(query = intersects, 
         params = list('da'),
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[1]],
         active = T),
    list(query = intersects, 
         params = list('da:da'),
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[1]],
         active = T),
    list(query = intersects, 
         params = list('scute'),
         color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[1]],
         active = T)
    ),
  # sets
  sets = names(up), keep.order= TRUE,
  sets.bar.color = cet_pal(n = 2, name = "cbd1", alpha = 1)[[1]],
  set_size.show = TRUE,
  set_size.numbers_size = 6,
  set_size.scale_max = ceiling(max(sapply( down , length)) / 100)*100,
  # intersections
  main.bar.color = 'gray50',
  group.by = 'sets',
  order.by = "degree",
  # labels
  number.angles = 0,
  mainbar.y.label = "Condition Intersections",
  sets.x.label = "DEGs per condition", 
  text.scale = c(1.5, 1.3,  # c(ylab, yticks,
                 1.5, 1.3,  #   xlab, xticks,
                 1.3, 1.3), #   setlabs, barnumbers)
  # connections
  matrix.color = 'gray50',
  shade.color = cet_pal(n = 5, name = "cbd1", alpha = 1)[[2]],
  point.size = 3,
  line.size = 0.75
)

pdf(file=paste0(figdir,'/UpSet_plot_down_logfc2.pdf'),
    width=8, height=4.5)
downsetplot
grid.text("Fold change ≤ ¼", x = 0.65, y=0.95, gp=gpar(fontsize=18)) # ½
dev.off()

downsetplot
grid.text("Set intersections for DEGs fc<¼", x = 0.65, y=0.95, gp=gpar(fontsize=18)) # ½
```
