==> 4_RNAseqVis_Venn.Rmd <==
---
title: "Visualisation of DESeq2 results with sets"
description: "DEG analysis based on DESeq2 and GSEA"
principal investigator: "Joaquín de Navascués"
researcher: "Aleix Puig, modified by Joaquín de Navascués"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
    theme: readable
    df_print: paged
---
```{r set-publication-theme, echo=FALSE, cache=FALSE}
ggplot2::theme_set(ggpubr::theme_pubr(base_size=10))
```

```{r setup, echo = FALSE, cache = FALSE}
knitr::opts_chunk$set(dev = 'png', 
                      fig.align = 'center', fig.height = 5, fig.width = 8.5, 
                      pdf.options(encoding = "ISOLatin9.enc"),
                      fig.path='integration/figures/', warning=FALSE, message=FALSE)
```

**Libraries needed:**
```{r load_libraries, warning=FALSE, echo=FALSE}
if (!require("librarian")) install.packages("librarian")
# data
librarian::shelf(dplyr, purrr, reshape2, biomaRt, DescTools, santoku)
# graphics
librarian::shelf(UpSetR, VennDiagram, ggvenn, ggVennDiagram, eulerr)
# convenience
librarian::shelf(here)
setwd(here::here()) # to distinguish from dplyr::here()
```

**Path to definitive images (outside repo):**
```{r define_dir2figs}
figdir <- paste0(c(head(str_split(getwd(),'/')[[1]],-1),
                   paste0(tail(str_split(getwd(),'/')[[1]],1), '_figures')),
                 collapse='/')
dir.create(figdir, showWarnings = FALSE)
```

# RNAseq results visualisation with sets

## 1 Getting ready

This gets us the DGE data from `DESeq2`, identified by FlyBase/Ensembl ID and gene symbol:
```{r load_DEG_data}
# experimental design and labels
targets <- readRDS('output/targets.RDS')
# DEG data
DaDaOE_deg <- readRDS('output/Control_vs_DaDaOE.RDS')
DaKD_deg <- readRDS('output/Control_vs_DaKD.RDS')
DaOE_deg <- readRDS('output/Control_vs_DaOE.RDS')
ScOE_deg <- readRDS('output/Control_vs_ScOE.RDS')
# gene IDs and symbols
ensembl = useEnsembl(biomart = "ENSEMBL_MART_ENSEMBL",
                     dataset="dmelanogaster_gene_ensembl",
                     host = "https://oct2022.archive.ensembl.org")
filters = listFilters(ensembl)       # defines filters for specific queries
attributes = listAttributes(ensembl) # defines features tro show
dlist <- getBM(attributes=c('ensembl_gene_id', 'external_gene_name'), mart = ensembl)
rownames(dlist) <- dlist$ensembl_gene_id
```

Now, for each condition, we will need to get the lists of misregulated genes for a given threshold of fold change:
```{r extract_regulated_sets, warning = FALSE}
# extract genes with abs(log2FC)
extract_regulated_sets <- function(list_of_degs, names_degs, fc_thresh=2) {
  # fc_thresh must be a positive number
  # list_of_degs is a list of dataframes
  # names_degs is a list of strings
  # they must have the same length
  reg_lvl <- paste0(' reg@log~2~FC≥', fc_thresh)
  breaks <- brk_manual(c(-fc_thresh, fc_thresh), left_vec = c(FALSE, TRUE))
  regulated_sets <- NULL
  for (l in 1:length(list_of_degs)) {
    degs_na <- dplyr::select(list_of_degs[[l]], c(log2FoldChange, padj, ensemblGeneID))
    # better not pass NAs to kiru
    degs <- na.omit(degs_na)
    # filter by log2FC threshold
    degs$reg <- 
      kiru(
        degs$log2FoldChange,
        breaks = breaks,
        extend = TRUE,
        labels=c("down", "non", "up")
      )
    # recover NAs as non-regulataed
    added_nas <- degs_na[!(rownames(degs_na) %in% rownames(na.omit(degs_na))),]
    added_nas$reg <- 'non'
    degs <- rbind(degs, added_nas)
    # filter by p-val
    degs$reg <- as.character(degs$reg)
    degs$reg <- ifelse(degs[x,'padj']<0.5, degs[x,'reg'], 'non')
    # store reg status
    #rownames(degs) <- degs$ensemblGeneID
    degs[ , names_degs[[l]] ] <- degs$reg
    regulated_sets[[l]] <- dplyr::select(degs, c(ensemblGeneID, names_degs[[l]]) )
  }
  regulated_sets <- purrr::reduce(regulated_sets, full_join, by='ensemblGeneID')
  
  return(regulated_sets)
}

list_of_degs <- list(DaKD_deg, DaOE_deg, DaDaOE_deg, ScOE_deg)
names_degs <- list(
  as.character(targets[targets$Condition=='DaKD','condition_md'][[1]]),
  as.character(targets[targets$Condition=='DaOE','condition_md'][[1]]),
  as.character(targets[targets$Condition=='DaDaOE','condition_md'][[1]]),
  as.character(targets[targets$Condition=='ScOE','condition_md'][[1]])
  )

regs <- extract_regulated_sets(list_of_degs, names_degs, fc_thresh=2)
```

## 2 Euler and Venn diagrams

https://cran.r-project.org/web/packages/eulerr/vignettes/introduction.html
https://github.com/yanlinlin82/ggvenn

```{r}

# extract genes with abs(log2FC)
define_regulation <- function(x, fc_thresh) {
  select_Significant_Genes <- x %>% 
    filter(padj <= 0.05) %>%
    filter(log2FoldChange %][% c(-fc_thresh, fc_thresh))
  
  select_Significant_Genes[, paste0('reg@abs(log2fc)>', fc_thresh)] <- cut(
    select_Significant_Genes$log2FoldChange, c(-Inf,0,Inf), c("DOWN", "UP")
    )
 # Only for adding gene names if necessary
   select_Significant_Genes <- merge(select_Significant_Genes, dlist, by=1, all.x = TRUE)
  return(select_Significant_Genes)
}

```








```{r}


upregulation <- function(select_Significant_Genes) {
  upregulation <- select_Significant_Genes %>% 
    filter(log2FoldChange >= 1) %>%
    dplyr::select(matches("ensembl|external"))
  upregulation$upregulation <- 1

  return(upregulation)
}

downregulation <- function(select_Significant_Genes) {
    downregulation <- select_Significant_Genes %>% 
      filter(log2FoldChange <= -1) %>%
      dplyr::select(matches("ensembl|external"))
    downregulation$downregulation <- 1
    
  return(downregulation)
}

# Create a data frame with the 4 datasets of regulation (da, Da:Da, sc and emcRNAi)
complete_regulation <- function(reg_da,reg_2D,reg_sc,reg_emc) {
  
  complete_regulation <- merge(reg_da, reg_2D, by=1, all=TRUE)
  complete_regulation <- merge(complete_regulation, reg_sc, by=1, all=TRUE)
  colnames(complete_regulation) <- c("ensembleGeneID","da","DaDa","sc")
  rownames(complete_regulation) <- complete_regulation$ensembleGeneID
  complete_regulation$ensembleGeneID <- NULL
  
  
  complete_regulation[is.na(complete_regulation)] <- 0
  
  return(complete_regulation)
}

likes <- function(samples, regu = complete_downregulation) {
  ppl <- regu
  names(ppl) <- c("d", "dd", "sc", "emc")
  for (i in 1:length(samples)) {
    ppl <- subset(ppl, ppl[samples[i]] == T)
  }
  nrow(ppl)
}

plotConditions <- function(a, ...) {
  grid.newpage()
  if (length(a) == 1) {
    out <- draw.single.venn(likes(a), ...)
  }
  if (length(a) == 2) {
    out <- draw.pairwise.venn(likes(a[1]), likes(a[2]), likes(a[1:2]), ...)
  }
  if (length(a) == 3) {
    out <- draw.triple.venn(likes(a[1]), likes(a[2]), likes(a[3]), likes(a[1:2]), 
                            likes(a[2:3]), likes(a[c(1, 3)]), likes(a), ...)
  }
  if (length(a) == 4) {
    out <- draw.quad.venn(likes(a[1]), likes(a[2]), likes(a[3]), likes(a[4]), 
                          likes(a[1:2]), likes(a[c(1, 3)]), likes(a[c(1, 4)]), likes(a[2:3]), 
                          likes(a[c(2, 4)]), likes(a[3:4]), likes(a[1:3]), likes(a[c(1, 2, 
                                                                                     4)]), likes(a[c(1, 3, 4)]), likes(a[2:4]), likes(a), ...)
  }
  if (!exists("out")) 
    out <- "Oops"
  return(out)
}



########################################################################

regulation_da <- define_regulation(WT_vs_da)
UPregulation_da <- upregulation(regulation_da)
DOWNregulation_da <- downregulation(regulation_da)
regulation_2D <- define_regulation(WT_vs_2D)
UPregulation_2D <- upregulation(regulation_2D)
DOWNregulation_2D <- downregulation(regulation_2D)
regulation_sc <- define_regulation(WT_vs_sc)
UPregulation_sc <- upregulation(regulation_sc)
DOWNregulation_sc <- downregulation(regulation_sc)


complete_upregulation <- complete_regulation(UPregulation_da[-2],UPregulation_2D[-2],UPregulation_sc[-2])
complete_downregulation <- complete_regulation(DOWNregulation_da[-2],DOWNregulation_2D[-2],DOWNregulation_sc[-2])

# Triple upregulation
draw.triple.venn(area1 = nrow(subset(complete_upregulation, da == 1)), area2 = nrow(subset(complete_upregulation, DaDa == 1)), area3 = nrow(subset(complete_upregulation, sc == 1)), n12 = nrow(subset(complete_upregulation, da == 1 & DaDa == 1)), n23 = nrow(subset(complete_upregulation, DaDa == 1 & sc == 1)), n13 = nrow(subset(complete_upregulation,da == 1 & sc == 1)), n123 = nrow(subset(complete_upregulation, da == 1 & DaDa == 1 & sc ==1)), category = c("Da > 2 fold UP", "DaDa > 2 fold UP", "Sc > 2 fold UP"), lty = "blank", fill = c("skyblue", "pink1", "mediumorchid"))
# Triple downregulation
draw.triple.venn(area1 = nrow(subset(complete_downregulation, da == 1)), area2 = nrow(subset(complete_downregulation, DaDa == 1)), area3 = nrow(subset(complete_downregulation, sc == 1)), n12 = nrow(subset(complete_downregulation, da == 1 & DaDa == 1)), n23 = nrow(subset(complete_downregulation, DaDa == 1 & sc == 1)), n13 = nrow(subset(complete_downregulation,da == 1 & sc == 1)), n123 = nrow(subset(complete_downregulation, da == 1 & DaDa == 1 & sc ==1)), category = c("Da > 2 fold DOWN", "DaDa > 2 fold DOWN", "Sc > 2 fold DOWN"), lty = "blank", fill = c("skyblue", "pink1", "mediumorchid"))

# Sc upregulation DaDa downregulation
scUP_dadaDOWN <- merge(UPregulation_sc, DOWNregulation_2D, by=1, all=TRUE)
colnames(scUP_dadaDOWN) <- c("ensembleGeneID", "scUP","DaDaDOWN")
rownames(scUP_dadaDOWN) <- scUP_dadaDOWN$ensembleGeneID
scUP_dadaDOWN$ensembleGeneID <- NULL

draw.pairwise.venn(area1 = nrow(subset(scUP_dadaDOWN, scUP == 1)), area2 = nrow(subset(scUP_dadaDOWN, DaDaDOWN == 1)),cross.area = nrow(subset(scUP_dadaDOWN, scUP == 1 & DaDaDOWN == 1)), category = c("sc > 2 fold UP", "Da:Da > 2 fold DOWN"), lty = rep("blank",2), fill = c("light blue", "mediumorchid"), alpha = rep(0.5, 2), cat.pos = c(0,0), cat.dist = rep(0.025, 2), scaled = FALSE)
# Sc downregulation DaDa upregulation
scDOWN_dadaUP <- merge(DOWNregulation_sc, UPregulation_2D, by=1, all=TRUE)
colnames(scDOWN_dadaUP) <- c("ensembleGeneID", "scDOWN","DaDaUP")
rownames(scDOWN_dadaUP) <- scDOWN_dadaUP$ensembleGeneID
scDOWN_dadaUP$ensembleGeneID <- NULL

draw.pairwise.venn(area1 = nrow(subset(scDOWN_dadaUP, scDOWN == 1)), area2 = nrow(subset(scDOWN_dadaUP, DaDaUP == 1)),cross.area = nrow(subset(scDOWN_dadaUP, scDOWN == 1 & DaDaUP == 1)), category = c("sc > 2 fold DOWN", "Da:Da > 2 fold UP"), lty = rep("blank",2), fill = c("light blue", "mediumorchid"), alpha = rep(0.5, 2), cat.pos = c(0,0), cat.dist = rep(0.025, 2), scaled = FALSE)

# All conditions upregulation (default in the function)
plotConditions(c("d", "dd", "sc", "emc"), category = c("Da > 2 fold UP", "DaDa > 2 fold UP","sc > 2 fold UP", "emcRNAi > 2 fold UP"), lty = "blank", fill = c("skyblue", "pink1", "mediumorchid", "orange"))

# All conditions downregulation (change in the function)
plotConditions(c("d", "dd", "sc", "emc"), category = c("Da > 2 fold DOWN", "DaDa > 2 fold DOWN","sc > 2 fold DOWN", "emcRNAi > 2 fold DOWN"), lty = "blank", fill = c("skyblue", "pink1", "mediumorchid", "orange"), scaled = TRUE)



####################################################

complete_upregulation_ex <- add_rownames(complete_downregulation, "name")
dSets <- melt(complete_upregulation_ex[-5], id = "name")
dSets <- (subset(dSets, value == 1))[1:2]
plot(venneuler(dSets))


####################################################

upset(complete_upregulation[-5], sets = c("da", "DaDa", "sc"), sets.bar.color = "#56B4E9",
      order.by = "freq", empty.intersections = "on")
grid.text("> 2 fold upregulation",x = 0.65, y=0.95, gp=gpar(fontsize=18))


upset(complete_downregulation[-5], sets = c("da", "DaDa", "sc"), sets.bar.color = "#56B4E9",
      order.by = "freq", empty.intersections = "on")
grid.text("> 2 fold downregulation",x = 0.65, y=0.95, gp=gpar(fontsize=18))
```
